% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VcfFileInstance.R
\name{VcfFileInstance}
\alias{VcfFileInstance}
\title{Exporting VCF File Instance}
\usage{
VcfFileInstance(filename)
}
\arguments{
\item{path}{}
}
\value{
a VCF file instance from \code{\link{readVcf}}
}
\description{
Exporting VCF File Instance
}
\examples{
VcfFileInstance(filename)
}
